Testing hook

+++ create_screen_profiles.sh
@@ -12,17 +12,21 @@ if [ ${#@} == 0 ]; then
 else
   output="source /home/$(whoami)/.screenrc\n"
   hostlist="./host_list"
-  type=$1
-  env=$2
+  if [ -e $hostlist ]; then
+    type=$1
+    env=$2
 
-  configfile="./.screenrc-${type}${env}";
+    configfile="./.screenrc-${type}${env}";
 
-  i=0;
-  for host in `cat ${hostlist} | grep -i "$type" | grep -i "$env"`;
-  do
-    shortname=`echo $host | cut -d'.' -f1`
-    output+="screen -t ${shortname} $((i++)) ssh $(whoami)@${host}\n"
-  done
-  echo -e "To execute screen -c ${configfile}:\n${output}";
-  echo -e $output > $configfile;
+    i=0;
+    for host in `cat ${hostlist} | grep -i "$type" | grep -i "$env"`;
+    do
+      shortname=`echo $host | cut -d'.' -f1`
+      output+="screen -t ${shortname} $((i++)) ssh $(whoami)@${host}\n"
+    done
+    echo -e $output > $configfile;
+    echo -e "${output}\n\nTo execute screen -c ${configfile}";
+  else
+    echo "Need to supply a host file"
+  fi;
 fi;
